MB-16087: Remove pre-generated PDF & man page files
No need to have pre-generated PDF & man page versions of the
memcached.json documentation in the repo - if required they can be
generated using the `memcached-manpage` build target.

Also make the INSTALL of the PDF / man page version optional, so `make
install` won't error if the user has a2x present but chooses not to
build the memcached-manpage target.

Change-Id: I2117c143ef5eb9a5cd77b654cf945d24df454adc