MB-20586: Update memcached to use cb_malloc memory API
For all instances where we control C-style memory allocation, replace
with calls to cbmalloc/cbrealloc etc.

There are a small number of instances where memory is allocated on our
behalf inside a library function, for example asprintf(). In such
cases we must still use free() to match the (internal) malloc().

Change-Id: I3e93d3a4bc2c8103e7da96ef448df1c226564528